iC4U3/requirements.txt
# IC4U requirements
# Usage: pip install -r requirements.txt

# For YoloV5 ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
tqdm>=4.64.0
ultralytics>=8.0.111
pandas>=1.1.4
seaborn>=0.11.0
# For IC4U Projects ------------------------------------------------------------------------
pyserial
googletrans==4.0.0-rc1
gtts
playsound
